I1M.BusquedaEnEscalada

Description

El patrón de búsqueda en escalada

Este módulo contiene la definición del patrón de búsqueda en escalada estudiado en el tema 15 del curso.

Además, en el tema se incluye como de casos de aplicación del patrón

el problema del cambio de monedas y

<http://bit.ly/1LIx5ui el algoritmo de Prim del mínimo árbol de expansión>.

Documentation

buscaEscalada :: Ord nodo => (nodo -> [nodo]) -> (nodo -> Bool) -> nodo -> [nodo]

(buscaEscalada s o e) es la lista de soluciones del problema de espacio de estado definido por la función sucesores (s), el objetivo (o) y el estado inicial (e), obtenidas buscando por escalada.