I1M.BusquedaEnEspaciosDeEstados

Description

El patrón de búsqueda en espacios de estados

Las características de los problemas de espacios de estados son:

un conjunto de las posibles situaciones o nodos que constituye el espacio de estados; estos son las potenciales soluciones que se necesitan explorar;

un conjunto de movimientos de un nodo a otros nodos, llamados los sucesores del nodo;

un nodo inicial;

un nodo objetivo, que es la solución.

Este módulo contiene la definición del patrón de búsqueda en espacios de estados estudiado en el tema 15 del curso. Además, en el tema se incluye dos casos de aplicación del patrón:

el problema de las n reinas y

el problema de la mochila.

Documentation

buscaEE :: Eq nodo => (nodo -> [nodo]) -> (nodo -> Bool) -> nodo -> [nodo]

(buscaEE s o e) es la lista de soluciones del problema de espacio de estado definido por la función sucesores (s), el objetivo (o) y el estado inicial (e).