I1M.BusquedaPrimeroElMejor

Description

El patrón de búsqueda por primero el mejor

Este módulo contiene la definición del patrón de búsqueda por primero el mejor estudiado en el tema 15 del curso.

Además, en el tema se incluye como de aplicación del patrón el problema del 8-puzzle.

Documentation

buscaPM :: Ord nodo => (nodo -> [nodo]) -> (nodo -> Bool) -> nodo -> [nodo]

(buscaPM s o e) es la lista de soluciones del problema de espacio de estado definido por la función sucesores (s), el objetivo (o) y el estado inicial (e), obtenidas buscando por primero el mejor.