I1M.Conjunto

Description

TAD (tipo abstracto de datos) de los conjuntos.

Este módulo contiene el código del TAD de los conjuntos estudiado en el tema 17 del curso.

Documentation

data Conj a

Tipo de dato de los conjuntos.

vacio :: Conj a

vacio es el conjunto vacío. Por ejemplo,

ghci> vacio
{}

esVacio :: Conj a -> Bool

(esVacio c) se verifica si c es el conjunto vacío. Por ejemplo,

λ> esVacio (foldr inserta vacio [2,5])
False
λ> esVacio vacio
True

pertenece :: Ord a => a -> Conj a -> Bool

(pertenece x c) se verifica si x pertenece al conjunto c. Por ejemplo,

λ> let c1 = foldr inserta vacio [2,5,3,2]
λ> pertenece 3 c1
True
λ> pertenece 4 c1
False

inserta :: Ord a => a -> Conj a -> Conj a

(inserta x c) es el conjunto obtenido añadiendo el elemento x al conjunto c. Por ejemplo,

λ> let c1 = foldr inserta vacio [2,5,3,2]
λ> c1
{2,3,5}
λ> inserta 3 c1
{2,3,5}
λ> inserta 4 c1
{2,3,4,5}

elimina :: Ord a => a -> Conj a -> Conj a

(elimina x c) es el conjunto obtenido eliminando el elemento x del conjunto c. Por ejemplo,

λ> let c1 = foldr inserta vacio [2,5,3,2]
λ> c1
{2,3,5}
λ> elimina 3 c1
{2,5}
λ> elimina 7 c1
{2,3,5}