I1M.Dinamica

Description

El patrón de programación dinámica

Este módulo contiene la definición del patrón de programación dinámica estudiado en el tema 15 del curso.

Además, en el tema se incluye como de casos de aplicación del patrón

la sucesión de Fibonacci,

el producto de cadenas de matrices,

los árboles binarios de búsqueda optimales,

<http://bit.ly/1IsuYDe los caminos mínimos entre todos los pares de nodos de un grafo> y

el problema del viajante.

Documentation

module I1M.Tabla

dinamica :: Ix i => (Tabla i v -> i -> v) -> (i, i) -> Tabla i v

(dinamica f r) wa la tabla de cálculo dinámica de la función f en el rango r.