I1M.Pila

Description

TAD (tipo abstracto de datos) de las pilas.

Este módulo contiene el código del TAD de las pilas estudiado en el tema 14 del curso.

Documentation

data Pila a

Tipo de dato de las pilas.

vacia :: Pila a

p1 es un ejemplo de pila que se usará en los siguientes ejemplos:

ghci> p1
1|2|3|-

p1 :: Pila Int p1 = apila 1 (apila 2 (apila 3 vacia))

vacia es la pila vacía. Por ejemplo,

  ghci> vacia
  -

apila :: a -> Pila a -> Pila a

(apila x p) es la pila obtenida añadiendo x encima de la pila p. Por ejemplo,

apila 4 (apila 1 (apila 2 (apila 3 vacia)))  ==  4|1|2|3|-

cima :: Pila a -> a

(cima p) es la cima de la pila p. Por ejemplo,

cima (apila 1 (apila 2 (apila 3 vacia)))  ==  1

desapila :: Pila a -> Pila a

(desapila p) es la pila obtenida suprimiendo la cima de la pila p. Por ejemplo,

desapila (apila 1 (apila 2 (apila 3 vacia)))  ==  2|3|-

esVacia :: Pila a -> Bool

(esVacia p) se verifica si p es la pila vacía. Por ejemplo,

esVacia (apila 1 (apila 2 (apila 3 vacia)))  ==  False
esVacia vacia                                ==  True