I1M.RecorridoEnProfundidad

Description

Recorrido de grafos en profundidad

En los ejemplos se usará el siguiente grafo

  +---> 2 <---+
  |           |
  |           |
  1 --> 3 --> 6 --> 5
  |                 |
  |                 |
  +---> 4 <---------+

definido por > g = creaGrafo D (1,6) > [(1,2,0),(1,3,0),(1,4,0),(3,6,0),(5,4,0),(6,2,0),(6,5,0)]

Documentation

recorridoEnProfundidad :: (Num p, Ix v) => v -> Grafo v p -> [v]

(recorridoEnProfundidad i g) es el recorrido en profundidad del grafo g desde el vértice i. Por ejemplo,

recorridoEnProfundidad 1 g  ==  [1,2,3,6,5,4]

recorridoEnProfundidad' :: (Num p, Ix v) => v -> Grafo v p -> [v]

(recorridoEnProfundidad' i g) es el recorrido en profundidad del grafo g desde el vértice i, usando la lista de los visitados como acumulador. Por ejemplo,

recorridoEnProfundidad' 1 g  ==  [1,2,3,6,5,4]